+++ borax/__init__.py
@@ -1,4 +1,4 @@
 # coding=utf8
 
-__version__ = '3.1.0'
+__version__ = '3.2.0'
 __author__ = 'kinegratii'