+++ borax/datasets/dict_datasets.py
@@ -1,9 +1,6 @@
 # coding=utf8
 
 
-from borax.datasets.join_ import join_one, join
-
-
 class DictDataset:
     def __init__(self, data, primary_field=None):
         self._data = []
@@ -18,18 +15,3 @@ class DictDataset:
     def __iter__(self):
         for item in self.data:
             yield item
-
-    def join(self, values, from_, to_, as_args=None, as_kwargs=None):
-        join(
-            self._data,
-            values=values,
-            from_=from_,
-            to_=to_,
-            as_args=as_args,
-            as_kwargs=as_kwargs,
-        )
-        return self
-
-    def join_one(self, values, from_, as_):
-        join_one(self._data, values=values, from_=from_, as_=as_)
-        return self