+++ borax/fetch.py
@@ -6,5 +6,5 @@ from borax.datasets.fetch import *  # noqa: F403
 
 warnings.warn(
     'This module is deprecated and will be removed in V3.3.Use borax.datasets.fetch instead.',
-    category=PendingDeprecationWarning
+    category=DeprecationWarning
 )